==> 1.data_create-Copy1.ipynb <==
import numpy as np
import pandas as pd

data_path='data/'
train_path=data_path+'AutoMaster_TrainSet.csv'

test_path=data_path+'AutoMaster_TestSet.csv'

ARTICLE_FILE = data_path+"train_text.txt"
SUMMARRY_FILE = data_path+"train_label.txt"

# import re


# def stopwordslist(stop_file='data/HIT_stop_words.txt'):
#     stopwords = [line.strip() for line in open(stop_file, encoding='UTF-8').readlines()]
#     return set(stopwords)


# def filter_sentence(sentence, stopwords):
#     def clean_sentence(sentence):
#         if isinstance(sentence, str):
#             return re.sub(
#                 r'[\s+\-\|\!\/\[\]\{\}_,.$%^*(+\"\')]+|[:：+——()?【】“”！，。？、~@#￥%……&*（）]+|车主说|技师说|语音|图片|你好|您好',
#                 '', sentence)
#         else:
#             return 'x'

#     sentence = clean_sentence(sentence)
#     words = jieba.cut(sentence)
#     # 去停用词，太长词也不行
#     ws = [w for w in words if w not in stopwords and len(w) < 10]
#     return ' '.join(ws)
# stop_words=stopwordslist()

train=pd.read_csv(train_path)
test=pd.read_csv(test_path)

test.head(2)

test=test.dropna()
test['merged'] = test[['Question', 'Dialogue']].apply(lambda x: ' '.join(x),axis=1)


test['lens']=test['merged'].apply(lambda x:len(x)).values

np.argmax(test['lens'].values)

test.iloc[8607]['merged']

# train['merged'] = train['merged'].apply(lambda x: filter_sentence(x,stop_words))

train['lens']=train['merged'].apply(lambda x:len(x)).values

s_train=train.sort_values(by=['lens'],ascending=False)

s_train['lens']

(s_train['lens']<1000).sum()

s_train['lens'].values.mean()

s_train['lens'].describe()

m_txt=s_train.iloc[1]['merged']

m_txt
